refactor: extract worker token capture host
