+++ nemo_rl/models/generation/vllm/vllm_worker_async.py
@@ -44,6 +44,7 @@ from nemo_rl.models.generation.interfaces import (
 from nemo_rl.models.generation.vllm.checkpoint_engine import (
     VllmAsyncCheckpointEngineRpcMixin,
 )
+from nemo_rl.models.generation.vllm.token_capture_host import TokenCaptureHost
 from nemo_rl.models.generation.vllm.utils import (
     attach_routed_experts_to_chat_response_choices,
     attach_token_information_to_chat_response_choices,
@@ -189,20 +190,7 @@ class VllmAsyncGenerationWorkerImpl(
         self._engine_loop = None
         self._http_engine_client = None
 
-        # Ledger-authoritative token capture (dormant until the
-        # setup_token_capture fan-out runs). The weight
-        # version is stamped per model call at begin_call time and rotated by
-        # the set_rollout_weight_version fan-out from the SC's _sync_weights.
-        self.token_capture = None
-        self._rollout_weight_version = 0
-        # In-flight captured calls keyed by id(request): (ActiveCall, the
-        # exact engine prompt ids recorded at preprocess time).
-        self._capture_calls: dict[int, tuple[Any, list[int]]] = {}
-        self._staging_source: Any | None = None
-        # Guarded by _prefix_cache_lock: _fetch_chain_prefix runs on executor
-        # threads (asyncio.to_thread), so lookups/evictions can be concurrent.
-        self._prefix_cache: dict[str, list[int]] = {}
-        self._prefix_cache_lock = threading.Lock()
+        self.capture_host = TokenCaptureHost()
 
         super().__init__(
             config,
@@ -436,243 +424,17 @@ class VllmAsyncGenerationWorkerImpl(
     async def report_dp_openai_server_base_url(self) -> Optional[str]:
         return self.base_url
 
-    def install_token_capture(self, capture: Any) -> None:
-        """Gym's ``install_capture`` seam (the ``CaptureHost`` contract)."""
-        self.token_capture = capture
-
     async def setup_token_capture(
         self, dp_cfg: dict[str, Any], staging_partition: str
     ) -> bool:
-        """Host ledger-authoritative token capture in this worker.
-
-        Fan-out target (token_capture.enabled only): builds the in-worker
-        data-plane client and TQTokenSink, then makes the single
-        ``install_capture`` call wiring Gym's engine-blind capture core +
-        vLLM adapter into this worker. Returns whether capture was installed
-        (False on non-model-owner ranks, which serve no HTTP).
-        """
+        """Install capture on model-owner ranks that serve HTTP requests."""
         if not self.is_model_owner:
             return False
-        # Deferred: nemo_gym is an optional extra absent in non-gym runs.
-        from nemo_gym.token_id_capture.adapters.vllm import VLLMCaptureAdapter
-        from nemo_gym.token_id_capture.staging import install_capture
-
-        from nemo_rl.data_plane import build_data_plane_client
-        from nemo_rl.data_plane.tq_token_sink import TQTokenSink, TQTokenSource
-
-        dp_client = build_data_plane_client(dp_cfg, bootstrap=False)
-        sink = TQTokenSink(dp_client, staging_partition=staging_partition)
-        self._staging_source = TQTokenSource(
-            dp_client, staging_partition=staging_partition
-        )
-        self._prefix_cache.clear()
-        install_capture(
-            self,
-            sink=sink,
-            weight_version_fn=lambda: self._rollout_weight_version,
-            adapter=VLLMCaptureAdapter(),
-        )
-        return True
+        return self.capture_host.setup(dp_cfg, staging_partition)
 
     async def set_rollout_weight_version(self, version: int) -> None:
-        """Rotate the weight version stamped on subsequent captured calls."""
-        self._rollout_weight_version = int(version)
-
-    def _capture_admission(self, request: Any) -> Any | None:
-        """Parse the ledger's ``ng_capture`` context into a ``CaptureAdmission``.
-
-        Returns None unless capture is installed and the request carries the
-        context. The dict itself is never mutated: the admission is the typed,
-        read-only contract that the prefix resolution and ``begin_call`` share.
-        """
-        context = getattr(request, "ng_capture", None)
-        if self.token_capture is None or not context:
-            return None
-        # Deferred: nemo_gym is an optional extra absent in non-gym runs.
-        from nemo_gym.token_id_capture.staging.records import CaptureAdmission
-
-        return CaptureAdmission.model_validate(context)
-
-    def _begin_request_capture(
-        self,
-        request: Any,
-        prompt_token_ids: list[int],
-        *,
-        admission: Any | None = None,
-        prefix_token_ids: list[int] | None = None,
-    ) -> None:
-        """Admit one ledger-forwarded call into the capture layer.
-
-        Called from preprocess_chat once the exact engine prompt is known
-        (post-splice in token-in mode, full render in text mode). No-op
-        unless capture is installed and the request carries the ledger's
-        ``ng_capture`` context.
-
-        ``prefix_token_ids`` is the prefix resolved by
-        :meth:`_resolve_admission_prefix`; Gym's ``begin_call`` checks it
-        against the admission (length == ``prev_len``, equal to an inline
-        prefix) and requires it for a ``staging_chain`` admission.
-        """
-        capture = self.token_capture
-        if capture is None:
-            return
-        if admission is None:
-            admission = self._capture_admission(request)
-            if admission is None:
-                return
-        call = capture.begin_call(
-            admission,
-            prefix_token_ids=prefix_token_ids,
-            stream=bool(getattr(request, "stream", False)),
-        )
-        self._capture_calls[id(request)] = (call, list(prompt_token_ids))
-
-    def _fetch_chain_prefix(self, staging_chain: list[str]) -> list[int]:
-        """Assemble prefix token ids from staging_chain, with a worker-local LRU cache."""
-        cache = self._prefix_cache
-        with self._prefix_cache_lock:
-            cached_ids: list[int] = []
-            miss_start = 0
-            for i, key in enumerate(staging_chain):
-                if key in cache:
-                    cached_ids = cache[key]
-                    miss_start = i + 1
-            miss_keys = staging_chain[miss_start:]
-        if not miss_keys:
-            return list(cached_ids)
-        if self._staging_source is None:
-            raise RuntimeError(
-                "_staging_source not initialized; call setup_token_capture() first"
-            )
-        # TQ read stays outside the lock so concurrent fetches overlap.
-        fetched = self._staging_source.fetch_prefix_token_ids(miss_keys)
-        result = cached_ids + fetched
-        last_key = staging_chain[-1]
-        with self._prefix_cache_lock:
-            cache[last_key] = result
-            if len(cache) > 256:
-                del cache[next(iter(cache))]
-        return result
-
-    def _resolve_admission_prefix(self, admission: Any) -> list[int]:
-        """Resolve a ``CaptureAdmission`` to the flat prefix the engine prompt starts with.
-
-        A ``staging_chain`` is fetched through the cached TransferQueue read;
-        an inline ``required_prefix_token_ids`` is used as is; a text root has
-        no prefix. Length checks are Gym's: ``begin_call`` rejects a prefix
-        that does not match ``prev_len``.
-        """
-        if admission.mode == "text":
-            return []
-        if admission.staging_chain:
-            return self._fetch_chain_prefix(list(admission.staging_chain))
-        return list(admission.required_prefix_token_ids)
-
-    def _enter_request_prefix(self, request: Any, prefix_token_ids: list[int]) -> None:
-        """Attach the resolved prefix to the request through the capture adapter.
-
-        ``VLLMCaptureAdapter.enter_prefix`` writes the engine-native field
-        (``required_prefix_token_ids``) into a payload; the same fields are
-        applied to the pydantic request so the existing prefix-splice branch
-        of preprocess_chat handles staged and inline prefixes alike.
-        """
-        adapter = self.token_capture.adapter
-        for field_name, value in adapter.enter_prefix({}, prefix_token_ids).items():
-            setattr(request, field_name, value)
-
-    @staticmethod
-    def _delta_align_routed_experts(
-        payload: dict[str, Any], *, prev_len: int, prompt_len: int, generated_len: int
-    ) -> None:
-        """Normalize optional vLLM routes to the exact staged token delta."""
-        choices = payload.get("choices") or []
-        if len(choices) != 1 or not isinstance(choices[0], dict):
-            return
-        choice = dict(choices[0])
-        message = dict(choice.get("message") or {})
-        routed = message.get("routed_experts")
-        if routed is None:
-            return
-        try:
-            from nemo_rl.utils.routed_experts_codec import (
-                decode_routed_experts,
-                encode_routed_experts,
-            )
-
-            if isinstance(routed, str):
-                dtype_name = routed.split(":", 3)[1]
-                dtype = {
-                    "int8": torch.int8,
-                    "int16": torch.int16,
-                    "int32": torch.int32,
-                }.get(dtype_name)
-                if dtype is None:
-                    raise ValueError(f"unsupported routed_experts dtype {dtype_name!r}")
-            else:
-                dtype = torch.int16
-            experts = decode_routed_experts(routed, dtype)
-            expected_full_len = prompt_len + generated_len
-            if experts.dim() != 3 or experts.shape[0] != expected_full_len:
-                raise ValueError(
-                    f"route length {experts.shape[0]} does not match engine sequence "
-                    f"length {expected_full_len}"
-                )
-            message["routed_experts"] = encode_routed_experts(experts[prev_len:])
-        except (IndexError, TypeError, ValueError) as error:
-            LOGGER.warning(
-                "dropping invalid routed_experts from staged capture: %s", error
-            )
-            message.pop("routed_experts", None)
-        choice["message"] = message
-        payload["choices"] = [choice]
-
-    def _finish_request_capture(self, request: Any, content: dict) -> dict:
-        """Stage the finished call and ride its coords on the response.
-
-        Fail-closed: the sink write happens inside complete_call —
-        the coords exist only after the bytes are durable, and any capture
-        failure degrades to capture_failed coords without breaking the
-        completion. Token ids and logprobs are stripped: the staged delta is
-        the only token store on this path, so the worker->gate hop carries
-        text + delta ids + coords only.
-        """
-        state = self._capture_calls.pop(id(request), None)
-        if state is None:
-            return content
-        call, prompt_token_ids = state
-        payload = dict(content)
-        # vLLM's OpenAI response carries no prompt ids; the adapter reads the
-        # preprocess-time engine prompt off the payload (see
-        # nemo_gym.token_id_capture.adapters.vllm.extract_prompt_ids).
-        payload["prompt_token_ids"] = prompt_token_ids
-        adapter = self.token_capture.adapter
-        if adapter is not None:
-            try:
-                generated_token_ids, _ = adapter.extract_generation(payload)
-            except Exception:  # capture core will report the authoritative failure
-                generated_token_ids = []
-            self._delta_align_routed_experts(
-                payload,
-                prev_len=call.admission.prev_len,
-                prompt_len=len(prompt_token_ids),
-                generated_len=len(generated_token_ids),
-            )
-        coords = self.token_capture.complete_call_from_response(call, payload)
-        for choice in content.get("choices") or []:
-            choice.pop("logprobs", None)
-            # The delta-aligned routes were staged to TQ above; the served
-            # full-length copy is dead weight the gate strips on arrival.
-            message = choice.get("message")
-            if isinstance(message, dict):
-                message.pop("routed_experts", None)
-        content["ng_commit_coords"] = coords.model_dump()
-        return content
-
-    def _abort_request_capture(self, request: Any, *, reason: str) -> None:
-        """Drop the in-flight capture state for a request that errored."""
-        state = self._capture_calls.pop(id(request), None)
-        if state is not None and self.token_capture is not None:
-            self.token_capture.fail_call(state[0], reason=reason)
+        """Rotate the version stamped on subsequent captured calls."""
+        self.capture_host.set_weight_version(version)
 
     # ruff: noqa
     def _setup_vllm_openai_api_server(self, app: FastAPI) -> FastAPI:
@@ -852,13 +614,15 @@ class VllmAsyncGenerationWorkerImpl(
                 # its event loop explicitly. The adapter then attaches the
                 # prefix to the request, so the inline-prefix branch below is
                 # the single splice path for staged and inline prefixes.
-                admission = worker_self._capture_admission(request)
+                admission = worker_self.capture_host.admission(request)
                 capture_prefix_token_ids: list[int] | None = None
                 if admission is not None and admission.mode == "token_in":
                     capture_prefix_token_ids = await asyncio.to_thread(
-                        worker_self._resolve_admission_prefix, admission
+                        worker_self.capture_host.resolve_prefix, admission
                     )
-                    worker_self._enter_request_prefix(request, capture_prefix_token_ids)
+                    worker_self.capture_host.enter_prefix(
+                        request, capture_prefix_token_ids
+                    )
 
                 if (
                     not hasattr(request, "required_prefix_token_ids")
@@ -873,7 +637,7 @@ class VllmAsyncGenerationWorkerImpl(
                         )
                     # Token capture, text mode: the full render is the exact
                     # engine prompt.
-                    worker_self._begin_request_capture(
+                    worker_self.capture_host.begin_request(
                         request, res[1][0]["prompt_token_ids"], admission=admission
                     )
                     return res
@@ -938,7 +702,7 @@ class VllmAsyncGenerationWorkerImpl(
                 # Token capture, token-in mode: the spliced prompt is the
                 # exact engine prompt; begin_call re-checks the prefix it
                 # was spliced from against the admission.
-                worker_self._begin_request_capture(
+                worker_self.capture_host.begin_request(
                     request,
                     final_prompt_token_ids,
                     admission=admission,
@@ -1159,7 +923,7 @@ class VllmAsyncGenerationWorkerImpl(
                 # max_model_len during tokenization, instead of returning an
                 # ErrorResponse. Convert to HTTP 400 so the Gym proxy can
                 # detect context-length overflow and handle it gracefully.
-                worker_self._abort_request_capture(request, reason="context_length")
+                worker_self.capture_host.abort_request(request, reason="context_length")
                 return JSONResponse(
                     content={
                         "error": {
@@ -1172,11 +936,11 @@ class VllmAsyncGenerationWorkerImpl(
                     status_code=400,
                 )
             except BaseException:
-                worker_self._abort_request_capture(request, reason="engine_error")
+                worker_self.capture_host.abort_request(request, reason="engine_error")
                 raise
 
             if isinstance(generator, ErrorResponse):
-                worker_self._abort_request_capture(request, reason="error_response")
+                worker_self.capture_host.abort_request(request, reason="error_response")
                 return JSONResponse(
                     content=generator.model_dump(), status_code=generator.error.code
                 )
@@ -1190,11 +954,11 @@ class VllmAsyncGenerationWorkerImpl(
                 # Off-loop: the sink write inside complete_call is a blocking
                 # TQ round trip (see the staging protocol's serving-host rule).
                 content = await asyncio.to_thread(
-                    worker_self._finish_request_capture, request, content
+                    worker_self.capture_host.finish_request, request, content
                 )
                 return JSONResponse(content=content)
 
-            worker_self._abort_request_capture(request, reason="streaming_response")
+            worker_self.capture_host.abort_request(request, reason="streaming_response")
             return StreamingResponse(content=generator, media_type="text/event-stream")
 
         ########################################
