+++ tests/unit/models/generation/test_vllm_token_capture_hosting.py
@@ -23,7 +23,6 @@ a mock worker group.
 from __future__ import annotations
 
 import asyncio
-import threading
 from types import SimpleNamespace
 from unittest.mock import MagicMock
 
@@ -46,6 +45,8 @@ from nemo_rl.models.generation.vllm.vllm_worker_async import (  # noqa: E402
     VllmAsyncGenerationWorkerImpl,
 )
 
+from nemo_rl.models.generation.vllm.token_capture_host import TokenCaptureHost
+
 pytestmark = pytest.mark.nemo_gym
 
 
@@ -60,18 +61,10 @@ class _MemorySink:
 
 def _fake_worker(*, is_model_owner: bool = True) -> SimpleNamespace:
     """The attribute surface setup_token_capture touches, minus the engine."""
-    worker = SimpleNamespace(
+    return SimpleNamespace(
         is_model_owner=is_model_owner,
-        token_capture=None,
-        _rollout_weight_version=0,
-        _staging_source=None,
-        _prefix_cache={},
-        _prefix_cache_lock=threading.Lock(),
+        capture_host=TokenCaptureHost(),
     )
-    worker.install_token_capture = lambda capture: setattr(
-        worker, "token_capture", capture
-    )
-    return worker
 
 
 def test_setup_token_capture_installs_capture_with_vllm_adapter(monkeypatch):
@@ -93,10 +86,10 @@ def test_setup_token_capture_installs_capture_with_vllm_adapter(monkeypatch):
     )
 
     assert installed is True
-    assert isinstance(worker.token_capture, RolloutTokenCapture)
-    assert worker.token_capture.adapter is not None
+    assert isinstance(worker.capture_host.token_capture, RolloutTokenCapture)
+    assert worker.capture_host.token_capture.adapter is not None
     # The adapter is the vLLM one (prefix ids enter via the worker's field).
-    payload = worker.token_capture.adapter.enter_prefix({}, [1, 2])
+    payload = worker.capture_host.token_capture.adapter.enter_prefix({}, [1, 2])
     assert payload["required_prefix_token_ids"] == [1, 2]
 
 
@@ -108,7 +101,7 @@ def test_setup_token_capture_skips_non_model_owners(monkeypatch):
         )
     )
     assert installed is False
-    assert worker.token_capture is None
+    assert worker.capture_host.token_capture is None
 
 
 def test_weight_version_is_stamped_from_worker_state(monkeypatch):
@@ -131,17 +124,17 @@ def test_weight_version_is_stamped_from_worker_state(monkeypatch):
     )
 
     asyncio.run(VllmAsyncGenerationWorkerImpl.set_rollout_weight_version(worker, 4))
-    first = worker.token_capture.begin_call(
+    first = worker.capture_host.token_capture.begin_call(
         CaptureAdmission(rollout_id="r", model_call_id="c1", mode="text")
     )
     asyncio.run(VllmAsyncGenerationWorkerImpl.set_rollout_weight_version(worker, 5))
-    second = worker.token_capture.begin_call(
+    second = worker.capture_host.token_capture.begin_call(
         CaptureAdmission(rollout_id="r", model_call_id="c2", mode="text")
     )
 
     assert (first.weight_version, second.weight_version) == (4, 5)
 
-    coords = worker.token_capture.complete_call(
+    coords = worker.capture_host.token_capture.complete_call(
         first, prompt_token_ids=[1], generated_token_ids=[2], generated_logprobs=[-0.1]
     )
     assert coords.weight_version == 4
@@ -203,23 +196,7 @@ class _FakeRequest(SimpleNamespace):
 def _worker_with_capture(sink: _MemorySink):
     from nemo_gym.token_id_capture.adapters.vllm import VLLMCaptureAdapter
 
-    worker = _fake_worker()
-    worker._capture_calls = {}
-    worker._prefix_cache = {}
-    worker._prefix_cache_lock = threading.Lock()
-    worker._staging_source = None
-    worker._delta_align_routed_experts = (
-        VllmAsyncGenerationWorkerImpl._delta_align_routed_experts
-    )
-    for name in (
-        "_fetch_chain_prefix",
-        "_capture_admission",
-        "_resolve_admission_prefix",
-        "_enter_request_prefix",
-    ):
-        setattr(
-            worker, name, getattr(VllmAsyncGenerationWorkerImpl, name).__get__(worker)
-        )
+    worker = TokenCaptureHost()
     worker.token_capture = RolloutTokenCapture(
         sink=sink,
         weight_version_fn=lambda: worker._rollout_weight_version,
@@ -268,13 +245,11 @@ def test_request_capture_round_trip_stages_and_rides_coords():
         },
         stream=False,
     )
-    VllmAsyncGenerationWorkerImpl._begin_request_capture(worker, request, [10, 11, 12])
+    TokenCaptureHost.begin_request(worker, request, [10, 11, 12])
     content = _served_content([13, 14], [-0.1, -0.2])
     # Full-length routes on the served response must not survive the strip.
     content["choices"][0]["message"]["routed_experts"] = [[[0]]] * 5
-    content = VllmAsyncGenerationWorkerImpl._finish_request_capture(
-        worker, request, content
-    )
+    content = TokenCaptureHost.finish_request(worker, request, content)
     # Bytes were staged before the coords existed (fail-closed ordering).
     assert len(sink.records) == 1
     assert sink.records[0].token_ids_delta == [10, 11, 12, 13, 14]
@@ -310,10 +285,8 @@ def test_request_capture_token_in_prev_len_chains():
         stream=False,
     )
     spliced_prompt = [10, 11, 12, 20, 21]  # exact prefix + fresh suffix
-    VllmAsyncGenerationWorkerImpl._begin_request_capture(
-        worker, request, spliced_prompt
-    )
-    content = VllmAsyncGenerationWorkerImpl._finish_request_capture(
+    TokenCaptureHost.begin_request(worker, request, spliced_prompt)
+    content = TokenCaptureHost.finish_request(
         worker, request, _served_content([22], [-0.5])
     )
     coords = content["ng_commit_coords"]
@@ -347,10 +320,10 @@ def test_staging_chain_prefix_flows_through_adapter_and_begin_call():
     request = _staging_chain_request()
     context_before = dict(request.ng_capture)
 
-    admission = worker._capture_admission(request)
-    prefix = worker._resolve_admission_prefix(admission)
-    worker._enter_request_prefix(request, prefix)
-    VllmAsyncGenerationWorkerImpl._begin_request_capture(
+    admission = worker.admission(request)
+    prefix = worker.resolve_prefix(admission)
+    worker.enter_prefix(request, prefix)
+    TokenCaptureHost.begin_request(
         worker,
         request,
         prefix + [20],
@@ -385,15 +358,15 @@ def test_inline_prefix_admission_resolves_without_a_fetch():
         },
         stream=False,
     )
-    admission = worker._capture_admission(request)
-    assert worker._resolve_admission_prefix(admission) == [10, 11]
+    admission = worker.admission(request)
+    assert worker.resolve_prefix(admission) == [10, 11]
     assert worker._staging_source.calls == []
-    text_root = worker._capture_admission(
+    text_root = worker.admission(
         _FakeRequest(
             ng_capture={"rollout_id": "r0", "model_call_id": "c1", "mode": "text"}
         )
     )
-    assert worker._resolve_admission_prefix(text_root) == []
+    assert worker.resolve_prefix(text_root) == []
 
 
 def test_staging_chain_cache_fetches_only_uncached_suffix():
@@ -401,10 +374,8 @@ def test_staging_chain_cache_fetches_only_uncached_suffix():
     source = _MemoryPrefixSource({"r0/c1": [10, 11], "r0/c2": [12]})
     worker._staging_source = source
 
-    first = VllmAsyncGenerationWorkerImpl._fetch_chain_prefix(worker, ["r0/c1"])
-    second = VllmAsyncGenerationWorkerImpl._fetch_chain_prefix(
-        worker, ["r0/c1", "r0/c2"]
-    )
+    first = TokenCaptureHost.fetch_chain_prefix(worker, ["r0/c1"])
+    second = TokenCaptureHost.fetch_chain_prefix(worker, ["r0/c1", "r0/c2"])
 
     assert first == [10, 11]
     assert second == [10, 11, 12]
@@ -418,11 +389,11 @@ def test_staging_chain_prefix_length_mismatch_is_rejected_by_begin_call():
     request = _staging_chain_request(prev_len=3)
     context_before = dict(request.ng_capture)
 
-    admission = worker._capture_admission(request)
-    prefix = worker._resolve_admission_prefix(admission)
+    admission = worker.admission(request)
+    prefix = worker.resolve_prefix(admission)
     assert prefix == [10, 11]
     with pytest.raises(CaptureError, match="does not equal prev_len 3"):
-        VllmAsyncGenerationWorkerImpl._begin_request_capture(
+        TokenCaptureHost.begin_request(
             worker, request, prefix + [20], admission=admission, prefix_token_ids=prefix
         )
 
@@ -435,9 +406,7 @@ def test_staging_chain_admission_requires_the_resolved_prefix_keyword():
     request = _staging_chain_request()
 
     with pytest.raises(CaptureError, match="pass the resolved prefix_token_ids"):
-        VllmAsyncGenerationWorkerImpl._begin_request_capture(
-            worker, request, [10, 11, 12, 20]
-        )
+        TokenCaptureHost.begin_request(worker, request, [10, 11, 12, 20])
 
     assert worker._capture_calls == {}
 
@@ -446,13 +415,11 @@ def test_request_capture_is_a_noop_without_context_or_capture():
     sink = _MemorySink()
     worker = _worker_with_capture(sink)
     plain = _FakeRequest(stream=False)  # no ng_capture attribute
-    VllmAsyncGenerationWorkerImpl._begin_request_capture(worker, plain, [1, 2])
+    TokenCaptureHost.begin_request(worker, plain, [1, 2])
     content = {
         "choices": [{"message": {"role": "assistant"}, "logprobs": {"content": []}}]
     }
-    out = VllmAsyncGenerationWorkerImpl._finish_request_capture(
-        worker, plain, dict(content)
-    )
+    out = TokenCaptureHost.finish_request(worker, plain, dict(content))
     assert "ng_commit_coords" not in out
     assert out["choices"][0]["logprobs"] is not None  # untouched off the capture path
     assert sink.records == []
@@ -471,14 +438,10 @@ def test_request_capture_abort_fails_the_call_and_drains_state():
         },
         stream=False,
     )
-    VllmAsyncGenerationWorkerImpl._begin_request_capture(worker, request, [1, 2])
-    VllmAsyncGenerationWorkerImpl._abort_request_capture(
-        worker, request, reason="engine_error"
-    )
+    TokenCaptureHost.begin_request(worker, request, [1, 2])
+    TokenCaptureHost.abort_request(worker, request, reason="engine_error")
     assert worker._capture_calls == {}
     assert sink.records == []
     # A late finish after abort is a no-op (state already drained).
-    out = VllmAsyncGenerationWorkerImpl._finish_request_capture(
-        worker, request, _served_content([3], [-0.1])
-    )
+    out = TokenCaptureHost.finish_request(worker, request, _served_content([3], [-0.1]))
     assert "ng_commit_coords" not in out
